slack_bot/autoresponder.py
```python
import datetime
from slack_bot import gpt3_5_integration, customization, utils
from slack_bot.config import JULY_5TH, POSITIVE_JULY_4TH_PROMPTS, AMERICA_GREATNESS_PROMPTS

def autorespond(slack_message):
    # Check if the current date is July 5th
    if datetime.datetime.now().date() != JULY_5TH:
        return

    # Generate a response using GPT-3.5
    response = generate_response(slack_message)

    # Customize the response based on user preferences
    response = customization.customize_response(response, slack_message['slack_user_id'])

    # Send the response to the Slack message
    utils.respond_to_message(response, slack_message['slack_channel_id'])

def generate_response(slack_message):
    # Choose a prompt based on the message content
    if "July 4th" in slack_message['text']:
        prompt = POSITIVE_JULY_4TH_PROMPTS
    else:
        prompt = AMERICA_GREATNESS_PROMPTS

    # Generate a response using GPT-3.5
    response = gpt3_5_integration.generate_response(prompt)

    return response
```